fix(api): strengthen route builder
With the previous version, url terminated by a slash were not supported. This
version rely on python's `urllib.parse.joinurl` function that is more resilient.

However, `urljoin` can only join 2 arguments. Moreover, it has a strange
behaviour:

```python
>>> urljoin("https://localhost/api/v2/assets", "1")
https://localhost/api/v2/1
```

As the last element in the first URL does not end with a `/`, it is overriden.

To avoid these problems, urljoin is wrapped in a `reduce` function.